Module thrust::reactor

Structs

Connection

FramedTransport
Wrap a TcpStream to handle reading and writing frames. Frames are simply some encoded thrift protocol byte buffer preceeded by a 32-bit unsigned length.

Id

Reactor
The Reactor is the component that interacts with networking. The reactor is built around Mio's event loop and manages both TcpListeners and TcpStreams.

Enums

Dispatch
Communication from the Reactor to outside components happens with a Dispatch message and normal Rust channels instead of Mio's variant.

FrameState

Message
Communication into the Mio event loop happens with a Message. For each new Mio event loop, a mio-specific Sender<Message> is returned.

State

Timeout